# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para buscar cep em diversos providers e retornar o mais rápido",
        "title": "Busca CEP",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Arthur Grigoletto",
            "url": "http://www.swagger.io/support",
            "email": "arthur.grigoletto.lima@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3333",
    "basePath": "/",
    "paths": {
        "/search/{cep}": {
            "get": {
                "description": "Get CEP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cep"
                ],
                "summary": "Get CEP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CEP",
                        "name": "cep",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Address"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        }
    }
}